Refactor asset system into modular packages
- split asset definitions into individual modules alongside helper and lifecycle layers\n- update docs, tests, and imports to use the new asset entry point\n\nManual testing: not run (CLI environment)\nAutomated testing: node --test tests/assetsFlow.test.js
