+++ src/game/assets/definitions/index.js
@@ -1,0 +1,17 @@
+import blogDefinition from './blog.js';
+import vlogDefinition from './vlog.js';
+import ebookDefinition from './ebook.js';
+import stockPhotosDefinition from './stockPhotos.js';
+import dropshippingDefinition from './dropshipping.js';
+import saasDefinition from './saas.js';
+
+const definitions = [
+  blogDefinition,
+  vlogDefinition,
+  ebookDefinition,
+  stockPhotosDefinition,
+  dropshippingDefinition,
+  saasDefinition
+];
+
+export default definitions;
