+++ src/game/assets/registry.js
@@ -1,0 +1,7 @@
+import definitions from './definitions/index.js';
+
+export const ASSETS = definitions;
+
+export default {
+  list: ASSETS
+};
